Remove the in-memory database from the development mode, as it is intended only for testing purposes.
