+++ internal/database/database-core.go
@@ -2,18 +2,16 @@ package database
 
 import (
 	"basement/main/internal/env"
-	"basement/main/internal/items"
 	"basement/main/internal/logg"
 	"database/sql"
 	"errors"
-	"fmt"
 	"os"
 
 	"github.com/gofrs/uuid/v5"
 	_ "modernc.org/sqlite"
 )
 
-const DATABASE_PROD_V1_FILE_PATH = "./internal/database/sqlite-database-prod-v1.db"
+const DATABASE_FILE_PATH = "./internal/database/sqlite-database.db"
 
 var ErrExist = errors.New("already exists")
 var ErrNotExist = errors.New("does not exist")
@@ -47,70 +45,48 @@ var triggers = &map[string]string{
 }
 
 type DB struct {
-	Sql *sql.DB
+	Sql       *sql.DB
+	fileExist bool
 }
 
 // Connect creates the database file if it doesn't exist and opens it.
 func (db *DB) Connect() {
-	if env.Development() {
-		logg.Info("using in-memory database")
-		db.open(":memory:")
-		db.InsertSampleItems()
-		db.InsertSampleBoxes()
-		db.InsertSampleShelves()
-		itemIDs, err := db.ItemIDs()
-		if err != nil {
-			logg.WrapErr(err)
-		}
-		boxIDs, err := db.BoxIDs()
-		if err != nil {
-			logg.WrapErr(err)
-		}
-		shelfRows, err := db.ShelfListRowsPaginated(1, 3)
-		if err != nil {
-			logg.WrapErr(err)
-		}
-		db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[0]), uuid.FromStringOrNil(boxIDs[0]))
-		db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[1]), uuid.FromStringOrNil(boxIDs[0]))
-		db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[2]), uuid.FromStringOrNil(boxIDs[0]))
-		db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[3]), uuid.FromStringOrNil(boxIDs[1]))
-		db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[4]), uuid.FromStringOrNil(boxIDs[1]))
-		db.MoveItemToShelf(uuid.FromStringOrNil(itemIDs[5]), shelfRows[0].ID)
-		db.MoveItemToShelf(uuid.FromStringOrNil(itemIDs[6]), shelfRows[0].ID)
-		db.MoveBoxToShelf(uuid.FromStringOrNil(boxIDs[0]), shelfRows[1].ID)
-		db.MoveBoxToShelf(uuid.FromStringOrNil(boxIDs[2]), shelfRows[1].ID)
-		db.MoveBoxToBox(uuid.FromStringOrNil(boxIDs[3]), uuid.FromStringOrNil(boxIDs[5]))
-		db.MoveBoxToBox(uuid.FromStringOrNil(boxIDs[3]), uuid.FromStringOrNil(boxIDs[5]))
-		db.MoveBoxToBox(uuid.FromStringOrNil(boxIDs[5]), uuid.FromStringOrNil(boxIDs[6]))
-	}
+	// create the database File and open it
+	db.createFile(DATABASE_FILE_PATH)
+	db.open(DATABASE_FILE_PATH)
 
-	if env.Production() {
-		logg.Infof(`using "%s" database`, DATABASE_PROD_V1_FILE_PATH)
-		db.createFile(DATABASE_PROD_V1_FILE_PATH)
-		db.open(DATABASE_PROD_V1_FILE_PATH)
+	// create the necessary Tables
+	db.createTable(*mainTables)
+	db.createTable(*virtualTables)
+	db.createTable(*triggers)
+
+	// add dummy data
+	if !db.fileExist && env.Development() {
+		db.insirtDummyData()
 	}
 }
 
 // createFile creates only db file if it doesn't exist, no tables.
-// If error occurs porgram will shut down with os.Exit(1).
+// If error occurs program will shut down with os.Exit(1).
 func (db *DB) createFile(dbFile string) {
 	_, err := os.Stat(dbFile)
 	if err == nil {
-		logg.Debugf(`"%s" exists`, dbFile)
+		db.fileExist = true
 		return
 	}
 
 	logg.Debugf(`creating "%s"`, dbFile)
-	file, err := os.Create("./sqlite-database-test.db")
+	file, err := os.Create(dbFile)
 	if err != nil {
 		logg.Fatalf("Failed to create database: %v", err)
 	}
 	defer file.Close()
 	logg.Infof(`"%s" created`, dbFile)
+
 }
 
 // open the connection and create tables if they don't exist.
-// If error occurs porgram will shut down with os.Exit(1).
+// If error occurs program will shut down with os.Exit(1).
 func (db *DB) open(dbFile string) {
 	var err error
 	db.Sql, err = sql.Open("sqlite", dbFile)
@@ -119,10 +95,6 @@ func (db *DB) open(dbFile string) {
 	}
 	logg.Debugf("opened '%s'", dbFile)
 	logg.Info("Database Connection established")
-
-	db.createTable(*mainTables)
-	db.createTable(*virtualTables)
-	db.createTable(*triggers)
 }
 
 func (db *DB) createTable(statements map[string]string) {
@@ -146,231 +118,32 @@ func (db *DB) ErrorExist() error {
 	return logg.WrapErrWithSkip(ErrExist, 2)
 }
 
-// Exists checks existence of entity (item, box, shelf, area).
-// Returns error if other internal errors happen.
-func (db *DB) Exists(entityType string, id uuid.UUID) (bool, error) {
-	err1 := ValidTable(entityType)
-	err2 := ValidVirtualTable(entityType)
-	// not a vaid table and not valid virtual table
-	if err1 != nil && err2 != nil {
-		return false, logg.NewError(fmt.Sprintf("no entity with type: \"%s\"", entityType))
-	}
-
-	var itemExists int
-	query := fmt.Sprintf(`SELECT EXISTS(SELECT 1 FROM %s WHERE id = ?)`, entityType)
-	err := db.Sql.QueryRow(query, id.String()).Scan(&itemExists)
+func (db *DB) insirtDummyData() {
+	db.InsertSampleItems()
+	db.InsertSampleBoxes()
+	db.InsertSampleShelves()
+	itemIDs, err := db.ItemIDs()
 	if err != nil {
-		if err == sql.ErrNoRows {
-			return false, nil
-		} else {
-			return false, logg.WrapErr(err)
-		}
+		logg.WrapErr(err)
 	}
-	if itemExists != 0 {
-		return true, nil
-	} else {
-		return false, nil
-	}
-}
-
-func (db *DB) deleteFrom(table string, id uuid.UUID) error {
-	err := ValidTable(table)
+	boxIDs, err := db.BoxIDs()
 	if err != nil {
-		return logg.WrapErr(err)
+		logg.WrapErr(err)
 	}
-
-	stmt := fmt.Sprintf(`DELETE FROM %s WHERE id = ?;`, table)
-	result, err := db.Sql.Exec(stmt, id.String())
+	shelfRows, err := db.ShelfListRowsPaginated(1, 3)
 	if err != nil {
-		return logg.Errorf(`can't delete "%s" from "%s" %w`, id, table, err)
+		logg.WrapErr(err)
 	}
-
-	rowsAffected, err := result.RowsAffected()
-	if err != nil {
-		return logg.WrapErr(err)
-	}
-	if rowsAffected == 0 {
-		return logg.NewError(fmt.Sprintf(`id "%s" not found in "%s"`, id, table))
-	} else if rowsAffected != 1 {
-		return logg.NewError(fmt.Sprintf(`unexpected number of rows affected (%d) while deleting "%s" from "%s"`, rowsAffected, id, table))
-	}
-	return nil
-}
-
-func ValidTable(table string) error {
-	_, ok := (*mainTables)[table]
-	if !ok {
-		return logg.NewError(fmt.Sprintf(`"%s" is not a valid table`, table))
-	}
-	return nil
-}
-
-func ValidVirtualTable(table string) error {
-	_, ok := (*virtualTables)[table]
-	if !ok {
-		return logg.NewError(fmt.Sprintf(`"%s" is not a valid virtual table`, table))
-	}
-	return nil
-}
-
-// MoveTo moves item/box/shelf to a box/shelf/area.
-//
-// Example move item to a box:
-//
-//	MoveTo("item", itemID, "box", boxID)
-//
-// To move things out set
-//
-//	toTableID = uuid.Nil
-func (db *DB) MoveTo(table string, id uuid.UUID, toTable string, toTableID uuid.UUID) error {
-	err := ValidTable(table)
-	if err != nil {
-		return logg.Errorf(`table: "%s" %w`, table, err)
-	}
-	err = ValidTable(toTable)
-	if err != nil {
-		return logg.Errorf(`toTable: "%s" %w`, toTable, err)
-	}
-
-	errMsg := fmt.Sprintf(`moving "%s" "%s" to "%s" "%s"`, table, id.String(), toTable, toTableID.String())
-
-	exists, err := db.Exists(table, id)
-	if err != nil {
-		return logg.WrapErr(err)
-	}
-	if exists == false {
-		return logg.Errorf("%s %w", errMsg, ErrNotExist)
-	}
-
-	// check if the table where the item is being moved to exists
-	if toTableID != uuid.Nil {
-		exists, err := db.Exists(toTable, toTableID)
-		if err != nil {
-			return logg.WrapErr(err)
-		}
-		if !exists {
-			return logg.Errorf("%s %w", errMsg, ErrNotExist)
-		}
-	}
-
-	// Update the item's shelf_id
-	stmt := fmt.Sprintf(`UPDATE %s SET %s_id = ? WHERE id = ?`, table, toTable)
-	result, err := db.Sql.Exec(stmt, toTableID.String(), id.String())
-	if err != nil {
-		return logg.WrapErr(err)
-	}
-	rows, err := result.RowsAffected()
-	if err != nil {
-		return logg.WrapErr(err)
-	}
-	if rows != 1 {
-		return logg.NewError(fmt.Sprintf("rows should be != 1 but is %d", rows))
-	}
-	return nil
-}
-
-// innerListRowsFrom returns all items/boxes/shelves/etc belonging to another box, shelf or area.
-//
-// Example:
-//
-//	// get all items that belongs to a shelf.
-//	innerListRowsFrom("shelf", shelf.ID, "item_fts")
-//
-// listRowsTable:
-//
-//	FROM "item_fts"
-//	FROM "box_ftx"
-//	...
-//
-// belongsToTable:
-//
-//	"item", "box", "shelf", ...
-//
-// belongsToTableID:
-//
-//	WHERE "item"_id = ID
-//	WHERE "box"_id = ID
-//	...
-func (db *DB) innerListRowsFrom(belongsToTable string, belongsToTableID uuid.UUID, listRowsTable string) ([]*items.ListRow, error) {
-	err := ValidVirtualTable(listRowsTable)
-	if err != nil {
-		return nil, logg.WrapErr(err)
-	}
-
-	err = ValidTable(belongsToTable)
-	if err != nil {
-		return nil, logg.WrapErr(err)
-	}
-
-	var listRows []*items.ListRow
-
-	stmt := fmt.Sprintf(`
-	SELECT id, label, box_id, box_label, shelf_id, shelf_label, area_id, area_label
-	FROM %s	
-	WHERE %s_id = ?;`, listRowsTable, belongsToTable)
-
-	rows, err := db.Sql.Query(stmt, belongsToTableID.String())
-	if err != nil {
-		return nil, logg.Errorf("%s %w", stmt, err)
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var sqlrow SQLListRow
-
-		err = rows.Scan(
-			&sqlrow.ID, &sqlrow.Label, &sqlrow.BoxID, &sqlrow.BoxLabel, &sqlrow.ShelfID, &sqlrow.ShelfLabel, &sqlrow.AreaID, &sqlrow.AreaLabel,
-		)
-		if err != nil {
-			return nil, logg.WrapErr(err)
-		}
-		lrow, err := sqlrow.ToListRow()
-		if err != nil {
-			return nil, logg.WrapErr(err)
-		}
-		listRows = append(listRows, lrow)
-	}
-	return listRows, nil
-}
-
-// Helper function to check for null strings and return empty if null
-func ifNullString(sqlStr sql.NullString) string {
-	if sqlStr.Valid {
-		return sqlStr.String
-	}
-	return ""
-}
-
-// Helper function to check for null strings and return empty if null
-func ifNullFloat64(sqlFloat sql.NullFloat64) float64 {
-	if sqlFloat.Valid {
-		return sqlFloat.Float64
-	}
-	return float64(0)
-}
-
-// Helper function to check for null strings and return empty if null
-func ifNullInt(sqlInt sql.NullInt64) int {
-	if sqlInt.Valid {
-		return int(sqlInt.Int64)
-	}
-	return 0
-}
-
-// Helper function to check for null UUIDs and return uuid.Nil if null
-func ifNullUUID(sqlUUID sql.NullString) uuid.UUID {
-	if sqlUUID.Valid {
-		return uuid.FromStringOrNil(sqlUUID.String)
-	}
-	return uuid.Nil
-}
-
-func UUIDFromSqlString(boxID sql.NullString) (uuid.UUID, error) {
-	if boxID.Valid {
-		id, err := uuid.FromString(boxID.String)
-		if err != nil {
-			return uuid.Nil, logg.Errorf("error while converting the string id into uuid: %w", err)
-		}
-		return id, nil
-	}
-	return uuid.Nil, logg.Errorf("invalid Virtual Id string")
+	db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[0]), uuid.FromStringOrNil(boxIDs[0]))
+	db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[1]), uuid.FromStringOrNil(boxIDs[0]))
+	db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[2]), uuid.FromStringOrNil(boxIDs[0]))
+	db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[3]), uuid.FromStringOrNil(boxIDs[1]))
+	db.MoveItemToBox(uuid.FromStringOrNil(itemIDs[4]), uuid.FromStringOrNil(boxIDs[1]))
+	db.MoveItemToShelf(uuid.FromStringOrNil(itemIDs[5]), shelfRows[0].ID)
+	db.MoveItemToShelf(uuid.FromStringOrNil(itemIDs[6]), shelfRows[0].ID)
+	db.MoveBoxToShelf(uuid.FromStringOrNil(boxIDs[0]), shelfRows[1].ID)
+	db.MoveBoxToShelf(uuid.FromStringOrNil(boxIDs[2]), shelfRows[1].ID)
+	db.MoveBoxToBox(uuid.FromStringOrNil(boxIDs[3]), uuid.FromStringOrNil(boxIDs[5]))
+	db.MoveBoxToBox(uuid.FromStringOrNil(boxIDs[3]), uuid.FromStringOrNil(boxIDs[5]))
+	db.MoveBoxToBox(uuid.FromStringOrNil(boxIDs[5]), uuid.FromStringOrNil(boxIDs[6]))
 }
